♻️ refactor: 목록추가 mutation분리
